Project now complete and ready for submission.  The binaries are compiled for 32-bit Linux.

+++ proxy.cpp
@@ -62,10 +62,9 @@ int main(int argc, char* argv[])
 
 void* client_thread(void* arg)
 {
-	//int csock = (int)arg;
-	long int csock = (long int)arg;
+	int csock = (int)arg;
 	
-	printf("[proxy] client ID #%ld connected\n", csock);
+	printf("[proxy] client ID #%d connected\n", csock);
 	
 	//create a new socket and connect to the bank
 	int bsock = socket(AF_INET, SOCK_STREAM, IPPROTO_TCP);
@@ -151,7 +150,7 @@ void* client_thread(void* arg)
 
 	}
 
-	printf("[proxy] client ID #%ld disconnected\n", csock);
+	printf("[proxy] client ID #%d disconnected\n", csock);
 		
 	close(bsock);
 	close(csock);
